PR fixes: change file names, make functions internal, change methods' responsibility
