+++ pkg/plugins/diff/tetetet.go
@@ -1,0 +1,1 @@
+package diff
