+++ pkg/plugins/wrapper.go
@@ -5,17 +5,17 @@ import (
 	"github.com/treeverse/lakefs/pkg/logging"
 )
 
-// PluginWrapper wraps the go-plugin client (which is in fact a plugin controller) and the GRPC client.
+// Wrapper wraps the go-plugin client (which is in fact a plugin controller) and the GRPC client.
 // It's used as a convenience struct to call the different types of plugins, and to be included as a subtype of specific
 // plugin client implementations
-type PluginWrapper struct {
+type Wrapper struct {
 	GRPCClient *plugin.ClientProtocol
 	Client     *plugin.Client
 	Log        logging.Logger
 }
 
 // InitClient initializes the inner GRPC client to communicate to the
-func (c *PluginWrapper) InitClient() error {
+func (c *Wrapper) InitClient() error {
 	if c.GRPCClient != nil {
 		return nil
 	}
@@ -29,7 +29,7 @@ func (c *PluginWrapper) InitClient() error {
 }
 
 // DestroyClient terminates the plugin
-func (c *PluginWrapper) DestroyClient() {
+func (c *Wrapper) DestroyClient() {
 	if c.Client != nil {
 		c.Client.Kill()
 	}
